Add dashboard HTML payload utility for resource dashboard
- Introduced `dashboard_html_payload.py` containing the `build_dashboard_payload` function to generate enriched dashboard data from test runs.
- Implemented logic for calculating CPU and RAM usage, as well as active time metrics, with options for chunk-based analysis.
- Integrated downsampling methods for optimizing data visualization in dashboards.
- Updated `tests_resource_dashboard.py` to utilize the new payload function, enhancing the resource dashboard's capabilities.
